# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation Taeyoung",
        "description": "API Documentation for Taeyoung",
        "version": "1.0.0"
    },
    "paths": {
...
        "/taeyoung-backend/public/api/sale": {
            "get": {
                "tags": [
                    "Sale"
                ],
                "summary": "List Sales",
                "description": "Returns a list of Sales.",
                "operationId": "d530e15bc7d7c298aef01ee02edba0ed",
                "parameters": [
                    {
                        "parameter": "page",
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "parameter": "per_page",
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "parameter": "documentType",
                        "name": "documentType",
                        "in": "query",
                        "description": "Document type",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "BOLETA",
                                "FACTURA"
                            ]
                        }
                    },
                    {
                        "parameter": "number",
                        "name": "number",
                        "in": "query",
                        "description": "Sale number",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "parameter": "date",
                        "name": "date[]",
                        "in": "query",
                        "description": "Sale date",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "format": "date"
                            }
                        }
                    },
                    {
                        "parameter": "client_id",
                        "name": "client_id",
                        "in": "query",
                        "description": "Client ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "parameter": "client$filterName",
                        "name": "client$filterName",
                        "in": "query",
                        "description": "Client name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "parameter": "client$country_id",
                        "name": "client$country_id",
                        "in": "query",
                        "description": "Client country ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "parameter": "quotation_id",
                        "name": "quotation_id",
                        "in": "query",
                        "description": "Quotation ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "parameter": "sort",
                        "name": "sort",
                        "in": "query",
                        "description": "Sort by column",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "parameter": "direction",
                        "name": "direction",
                        "in": "query",
                        "description": "Sort direction",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Client"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthenticated"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Sale"
                ],
                "summary": "Store Sale",
                "description": "Stores a Sale.",
                "operationId": "8d483a26901987969ee3fd0a21efd07f",
                "requestBody": {
                    "description": "Sale data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreSaleRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SaleResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ValidationError": {
                "properties": {
                    "error": {
                        "type": "string",
                        "example": "The pagination must be an integer."
                    }
                },
                "type": "object"
            },
            "Unauthenticated": {
                "properties": {
                    "error": {
                        "type": "string",
                        "example": "Unauthenticated."
                    }
                },
                "type": "object"
            },
...
            "StoreSaleRequest": {
                "title": "StoreSaleRequest",
                "required": [
                    "date",
                    "documentType",
                    "paymentType",
                    "quotas",
                    "client_id",
                    "detailMachinery|detailSpareParts"
                ],
                "properties": {
                    "date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-08-04"
                    },
                    "documentType": {
                        "type": "string",
                        "enum": [
                            "BOLETA",
                            "FACTURA"
                        ],
                        "example": "BOLETA"
                    },
                    "paymentType": {
                        "type": "string",
                        "enum": [
                            "CONTADO",
                            "CREDITO"
                        ],
                        "example": "CREDITO"
                    },
                    "quotation_id": {
                        "type": "integer",
                        "example": "21"
                    },
                    "client_id": {
                        "type": "integer",
                        "example": "21"
                    },
                    "currencyType": {
                        "type": "string",
                        "example": "PEN"
                    },
                    "discount": {
                        "type": "number",
                        "example": "100"
                    },
                    "quotas": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "days": {
                                    "type": "integer",
                                    "example": "7"
                                },
                                "amount": {
                                    "type": "number",
                                    "example": "14378.6"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "detailMachinery": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "description": {
                                    "type": "string",
                                    "example": "Machinery description"
                                },
                                "quantity": {
                                    "type": "integer",
                                    "example": "2"
                                },
                                "salePrice": {
                                    "type": "number",
                                    "example": "4000"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "detailSpareParts": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "quantity": {
                                    "type": "integer",
                                    "example": "4"
                                },
                                "salePrice": {
                                    "type": "number",
                                    "example": "767.5"
                                },
                                "spare_part_id": {
                                    "type": "integer",
                                    "example": "1"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
            "Client": {
                "required": [
                    "id",
                    "ruc",
                    "businessName",
                    "email",
                    "phone",
                    "country"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "type": {
                        "type": "string",
                        "example": "RUC"
                    },
                    "dni": {
                        "type": "string",
                        "example": "12345678"
                    },
                    "ruc": {
                        "type": "string",
                        "example": "12345678901"
                    },
                    "names": {
                        "type": "string",
                        "example": "Names"
                    },
                    "fatherSurname": {
                        "type": "string",
                        "example": "Father Surname"
                    },
                    "motherSurname": {
                        "type": "string",
                        "example": "Mother Surname"
                    },
                    "businessName": {
                        "type": "string",
                        "example": "Business Name"
                    },
                    "filterName": {
                        "type": "string",
                        "example": "Filter Name"
                    },
                    "address": {
                        "type": "string",
                        "example": "Address"
                    },
                    "email": {
                        "type": "string",
                        "example": "mail@gmail.com"
                    },
                    "phone": {
                        "type": "integer",
                        "example": "123456789"
                    },
                    "representativeDni": {
                        "type": "string",
                        "example": "12345678"
                    },
                    "representativeNames": {
                        "type": "string",
                        "example": "Representative Names"
                    },
                    "country": {
                        "type": "string",
                        "example": "Peru"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}